requirements.txt
azure-functions
sendgrid>=6.11.0
python-dotenv>=1.0.0
email-validator>=2.0.0
